requirements.txt
Flask~=2.0.3
Pillow~=9.0.1
opencv-python~=4.5.5.64